tracks streaming

+++ SolanaMusicApi.Application/Services/BaseService/IBaseService.cs
@@ -3,12 +3,12 @@
 public interface IBaseService<T>
 {
     IQueryable<T> GetAll();
-    Task<T> GetByIdAsync(int id);
+    Task<T> GetByIdAsync(long id);
     Task<T> Add(T entity);
     Task<IEnumerable<T>> AddRangeAsync(IEnumerable<T> entities);
     Task UpdateAsync(T entity);
     Task UpdateRangeAsync(IEnumerable<T> entities);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
     Task DeleteRangeAsync(IEnumerable<T> entities);
 
     Task BeginTransactionAsync();
